fix: Set icon for Android to avoid crash
The library that react-native-navigation currently uses,
https://github.com/aurelhubert/ahbottomnavigation, is meant for
Material Design, which requires an icon to be present. See
the related issue:
https://github.com/wix/react-native-navigation/issues/4684

The spec for this project does not have an icon so on iOS an
icon is not specified, but on Android there's an SVG file that
avoids the crash, but doesn't render to the screen. The alignment
of the text is off, but only slightly so I'm letting that slide
for now.

Also added several defaults to the navigation styling to start out