feat: Add RN CLI config and Inter fonts
The font in the spec file appears to be SF UI, the default iOS
font, but that is licensed to only be used in mockups unless
it's an official Apple product. I personally like Inter a lot
and to me appears to be somewhat visually similar to the spec.

The fonts were automatically linked running the command:
`react-native link`